feat: enforce accept json header on requests
This is a JSON API, so we always want to return JSON responses. However,
the client might not always explicitly request this. This middleware
sets the appropriate header to tell the framework to always return JSON
so that, for example, validation failures come back as a JSON response
with the validation messages in the body, as opposed to redirecting as
is the default behaviour for non-JSON requests.